Inline `sanitize_messages` into `messages`, drop standalone `sanitization` module
Rename the new public `sanitize_message_history` to `sanitize_messages` and move
it (plus its private helpers) into `messages.py` next to `narrow_message_parts`,
its sibling standalone function over a message list. Expose it only from
`pydantic_ai.messages`, matching `narrow_message_parts`/`is_multi_modal_content`,
rather than from the top-level `pydantic_ai` namespace.

Remove the `pydantic_ai.sanitization` module and its `api/sanitization.md`
reference page; the function now renders under `api/messages.md`. Repoint the
docs links and rename `tests/test_sanitization.py` accordingly.
