--- docs/api/sanitization.md
@@ -1,3 +1,0 @@
-# `pydantic_ai.sanitization`
-
-::: pydantic_ai.sanitization
