+++ docs/message-history.md
@@ -315,16 +315,17 @@ _(This example is complete, it can be run "as is")_
 
 The `message_history` parameter is trusted server-side state. If you load history that came from a browser request or another untrusted boundary, sanitize it before passing it to the agent.
 
-[`sanitize_message_history`][pydantic_ai.sanitization.sanitize_message_history] applies the same default message sanitization used by the [UI adapters](ui/overview.md): it strips client-supplied system prompts, drops non-HTTP file URL schemes, resets non-allowlisted [`FileUrl.force_download`][pydantic_ai.messages.FileUrl.force_download] values to `False`, drops uploaded file references, and removes unresolved tool calls at the end of the history.
+[`sanitize_messages`][pydantic_ai.messages.sanitize_messages] applies the same default message sanitization used by the [UI adapters](ui/overview.md): it strips client-supplied system prompts, drops non-HTTP file URL schemes, resets non-allowlisted [`FileUrl.force_download`][pydantic_ai.messages.FileUrl.force_download] values to `False`, drops uploaded file references, and removes unresolved tool calls at the end of the history.
 
 ```python {title="sanitize untrusted message history" test="skip" lint="skip"}
-from pydantic_ai import Agent, ModelMessagesTypeAdapter, sanitize_message_history
+from pydantic_ai import Agent, ModelMessagesTypeAdapter
+from pydantic_ai.messages import sanitize_messages
 
 agent = Agent('openai:gpt-5.2', instructions='Be a helpful assistant.')
 
 # `request_json` is the body submitted by an untrusted client.
 loaded_history = ModelMessagesTypeAdapter.validate_python(request_json['message_history'])
-message_history = sanitize_message_history(loaded_history)
+message_history = sanitize_messages(loaded_history)
 
 result = agent.run_sync('Tell me a different joke.', message_history=message_history)
 ```
