+++ pydantic_ai_slim/pydantic_ai/__init__.py
@@ -130,7 +130,6 @@ from .profiles import (
     ModelProfileSpec,
 )
 from .run import AgentRun, AgentRunResult, AgentRunResultEvent
-from .sanitization import sanitize_message_history
 from .settings import ModelSettings, ToolChoice, ToolOrOutput
 from .tools import (
     AgentNativeTool,
@@ -266,7 +265,6 @@ __all__ = (
     'VideoFormat',
     'VideoMediaType',
     'VideoUrl',
-    'sanitize_message_history',
     # profiles
     'ModelProfile',
     'ModelProfileSpec',
