+++ tests/test_sanitize_messages.py
@@ -15,14 +15,14 @@ from pydantic_ai import (
     ToolReturnPart,
     UploadedFile,
     UserPromptPart,
-    sanitize_message_history,
 )
+from pydantic_ai.messages import sanitize_messages
 
 from ._inline_snapshot import snapshot
 from .conftest import IsDatetime
 
 
-def test_sanitize_message_history_resets_force_download_from_serialized_history():
+def test_sanitize_messages_resets_force_download_from_serialized_history():
     serialized = [
         {
             'parts': [
@@ -44,7 +44,7 @@ def test_sanitize_message_history_resets_force_download_from_serialized_history(
     messages = ModelMessagesTypeAdapter.validate_python(serialized)
 
     with pytest.warns(UserWarning, match=r'force_download.*allow-local.*reset to `False`'):
-        sanitized = sanitize_message_history(messages)
+        sanitized = sanitize_messages(messages)
 
     message = sanitized[0]
     assert isinstance(message, ModelRequest)
@@ -58,21 +58,21 @@ def test_sanitize_message_history_resets_force_download_from_serialized_history(
     )
 
 
-def test_sanitize_message_history_keeps_resolved_trailing_tool_call():
+def test_sanitize_messages_keeps_resolved_trailing_tool_call():
     messages: list[ModelMessage] = [
         ModelRequest(parts=[UserPromptPart(content='do the thing')]),
         ModelResponse(parts=[ToolCallPart(tool_name='do_thing', tool_call_id='call-1')]),
     ]
 
-    kept = sanitize_message_history(messages, resolved_tool_call_ids=['call-1'])
+    kept = sanitize_messages(messages, resolved_tool_call_ids=['call-1'])
     assert kept == snapshot(messages)
 
     with pytest.warns(UserWarning, match=r'unresolved tool call.*do_thing'):
-        dropped = sanitize_message_history(messages)
+        dropped = sanitize_messages(messages)
     assert dropped == snapshot([ModelRequest(parts=[UserPromptPart(content='do the thing', timestamp=IsDatetime())])])
 
 
-def test_sanitize_message_history_strips_dangling_call_exposed_by_dropped_tail():
+def test_sanitize_messages_strips_dangling_call_exposed_by_dropped_tail():
     """A dangling tool call re-exposed as the tail by a dropped trailing message is still stripped.
 
     A trailing system prompt sanitizes to an empty `ModelRequest` that is dropped, promoting a
@@ -87,12 +87,12 @@ def test_sanitize_message_history_strips_dangling_call_exposed_by_dropped_tail()
 
     # Two warnings fire here (system-prompt strip + dangling call); assert on the dangling one.
     with pytest.warns(UserWarning) as record:
-        sanitized = sanitize_message_history(messages)
+        sanitized = sanitize_messages(messages)
     assert sanitized == []
     assert any('unresolved tool call' in str(w.message) and 'delete_account' in str(w.message) for w in record)
 
 
-def test_sanitize_message_history_keeps_resolved_call_exposed_by_dropped_tail():
+def test_sanitize_messages_keeps_resolved_call_exposed_by_dropped_tail():
     """A *resolved* tool call re-exposed as the tail by a dropped trailing message is kept.
 
     Same shape as the dropped-tail case above, but the trailing call is in
@@ -105,11 +105,11 @@ def test_sanitize_message_history_keeps_resolved_call_exposed_by_dropped_tail():
     ]
 
     with pytest.warns(UserWarning, match=r'system prompts were stripped'):
-        sanitized = sanitize_message_history(messages, resolved_tool_call_ids=['call-1'])
+        sanitized = sanitize_messages(messages, resolved_tool_call_ids=['call-1'])
     assert sanitized == [messages[0]]
 
 
-def test_sanitize_message_history_strips_dangling_call_but_keeps_other_tail_parts():
+def test_sanitize_messages_strips_dangling_call_but_keeps_other_tail_parts():
     """When the surviving tail response mixes a dangling call with other parts, only the call is stripped.
 
     The injected `ToolCallPart` is removed so a promptless run can't dispatch it, but the response's
@@ -123,23 +123,23 @@ def test_sanitize_message_history_strips_dangling_call_but_keeps_other_tail_part
     ]
 
     with pytest.warns(UserWarning, match=r'unresolved tool call.*delete_account'):
-        sanitized = sanitize_message_history(messages)
+        sanitized = sanitize_messages(messages)
 
     tail = sanitized[-1]
     assert isinstance(tail, ModelResponse)
     assert tail.parts == [messages[1].parts[0]]
 
 
-def test_sanitize_message_history_drops_empty_response():
+def test_sanitize_messages_drops_empty_response():
     """An empty-parts `ModelResponse` in untrusted history is dropped, not kept as `parts=[]`."""
     messages: list[ModelMessage] = [
         ModelRequest(parts=[UserPromptPart(content='hi')]),
         ModelResponse(parts=[]),
     ]
-    assert sanitize_message_history(messages) == [messages[0]]
+    assert sanitize_messages(messages) == [messages[0]]
 
 
-def test_sanitize_message_history_keeps_bytearray_tool_return_content():
+def test_sanitize_messages_keeps_bytearray_tool_return_content():
     """A `bytearray` tool return must be left intact, not iterated into a list of ints.
 
     `bytearray` is a `Sequence`, so the recursive tool-return walker has to exclude it alongside
@@ -150,7 +150,7 @@ def test_sanitize_message_history_keeps_bytearray_tool_return_content():
         ModelResponse(parts=[ToolCallPart(tool_name='read_bytes', tool_call_id='call-1')]),
         ModelRequest(parts=[ToolReturnPart(tool_name='read_bytes', content=bytearray(b'abc'), tool_call_id='call-1')]),
     ]
-    sanitized = sanitize_message_history(messages, resolved_tool_call_ids=['call-1'])
+    sanitized = sanitize_messages(messages, resolved_tool_call_ids=['call-1'])
     request = sanitized[1]
     assert isinstance(request, ModelRequest)
     part = request.parts[0]
@@ -158,7 +158,7 @@ def test_sanitize_message_history_keeps_bytearray_tool_return_content():
     assert part.content == bytearray(b'abc')
 
 
-def test_sanitize_message_history_strips_client_system_prompts():
+def test_sanitize_messages_strips_client_system_prompts():
     """Client-submitted system prompts are stripped by default (`strip_system_prompts=True`).
 
     The system prompt is the server's to own; a client that can inject one can override the agent's
@@ -169,18 +169,18 @@ def test_sanitize_message_history_strips_client_system_prompts():
     ]
 
     with pytest.warns(UserWarning, match=r'Client-submitted system prompts were stripped'):
-        sanitized = sanitize_message_history(messages)
+        sanitized = sanitize_messages(messages)
     request = sanitized[0]
     assert isinstance(request, ModelRequest)
     assert [type(p).__name__ for p in request.parts] == snapshot(['UserPromptPart'])
 
-    kept = sanitize_message_history(messages, strip_system_prompts=False)
+    kept = sanitize_messages(messages, strip_system_prompts=False)
     request = kept[0]
     assert isinstance(request, ModelRequest)
     assert [type(p).__name__ for p in request.parts] == snapshot(['SystemPromptPart', 'UserPromptPart'])
 
 
-def test_sanitize_message_history_drops_non_http_file_url_schemes():
+def test_sanitize_messages_drops_non_http_file_url_schemes():
     """Non-HTTP `FileUrl` schemes are dropped by default.
 
     A scheme like `s3://` is fetched by the provider with the server-side IAM role, so an untrusted
@@ -201,7 +201,7 @@ def test_sanitize_message_history_drops_non_http_file_url_schemes():
     ]
 
     with pytest.warns(UserWarning, match=r"scheme\(s\) \['s3'\] were dropped"):
-        sanitized = sanitize_message_history(messages)
+        sanitized = sanitize_messages(messages)
     request = sanitized[0]
     assert isinstance(request, ModelRequest)
     part = request.parts[0]
@@ -209,7 +209,7 @@ def test_sanitize_message_history_drops_non_http_file_url_schemes():
     assert part.content == snapshot(['look at this', ImageUrl(url='https://example.com/ok.png')])
 
 
-def test_sanitize_message_history_drops_uploaded_files_by_default():
+def test_sanitize_messages_drops_uploaded_files_by_default():
     """Client-submitted `UploadedFile`s are dropped unless `allow_uploaded_files=True`.
 
     Like a non-HTTP file URL, an uploaded file references an object the provider fetches with the
@@ -229,14 +229,14 @@ def test_sanitize_message_history_drops_uploaded_files_by_default():
     ]
 
     with pytest.warns(UserWarning, match=r"uploaded file\(s\) for provider\(s\) \['openai'\] were dropped"):
-        sanitized = sanitize_message_history(messages)
+        sanitized = sanitize_messages(messages)
     request = sanitized[0]
     assert isinstance(request, ModelRequest)
     part = request.parts[0]
     assert isinstance(part, UserPromptPart)
     assert part.content == snapshot(['summarize'])
 
-    kept = sanitize_message_history(messages, allow_uploaded_files=True)
+    kept = sanitize_messages(messages, allow_uploaded_files=True)
     request = kept[0]
     assert isinstance(request, ModelRequest)
     part = request.parts[0]
